refactor: 참가자의 이름 입력에 대해 관리하는 Names 추가
