+++ README.md
@@ -67,14 +67,15 @@
 
 - [x] 게스트들 
   - [x] 참여자의 목록을 가진다
+  - [x] 입력된 참가자 목록의 수와 베팅 금액 목록의 수가 다르면 예외를 발생한다.
 
 - [x] 이름
   - [x] 이름은 공백일 수 없다.
   - [x] 양쪽 공백을 제거한 이름은 1글자 이상, 15글자 이하여야 한다.
 
-- [ ] 이름들
-  - [ ] 참여자들의 이름은 1명 이상 10명 이하이다.
-  - [ ] 참여자들의 이름은 중복될 수 없다.
+- [x] 이름들
+  - [x] 참여자들의 이름은 1명 이상 10명 이하이다.
+  - [x] 참여자들의 이름은 중복될 수 없다.
   
 - [x] 게임 결과
   - [x] 플레이어들을 조회하여 게임결과를 취합한다.
